+++ package.json
@@ -33,6 +33,7 @@
         "typescript": "^5.5.3"
     },
     "dependencies": {
+        "joi": "^17.13.3",
         "redis": "^4.6.15"
     }
 }
