+++ src/index.ts
@@ -1,4 +1,4 @@
 import { setupRedisClient } from "./config/redisClient";
-import rateLimiter from "./middleware/rateLimiter";
+import rateLimiter from "./middleware/slidingWindow/rate-limiter";
 
 export {setupRedisClient, rateLimiter};
