+++ src/utils/errors/api-errors/NotAcceptableError.ts
@@ -1,0 +1,10 @@
+import { CustomApiError } from './CustomApiError';
+
+class NotAcceptableError extends CustomApiError {
+  statusCode = 406;
+  constructor(message: string) {
+    super(message);
+  }
+}
+
+export { NotAcceptableError };
